Refactor vault management by moving vault-related functions to a dedicated vaults storage module, simplifying the core storage interface. Remove obsolete vault functions from the main storage file and update references throughout the extension and desktop clients.
