+++ core/ui/storage/vaults.tsx
@@ -11,8 +11,29 @@ import { useMutation, useQuery } from '@tanstack/react-query'
 import { useMemo } from 'react'
 
 import { getVaultId, Vault } from '../vault/Vault'
-import { DeleteVaultFunction } from './CoreStorage'
 import { StorageKey } from './StorageKey'
+
+export type UpdateVaultInput = {
+  vaultId: string
+  fields: Partial<Vault>
+}
+
+export const vaultsInitialValue: Vault[] = []
+
+export type CreateVaultFunction = (vault: Vault) => Promise<Vault>
+
+export type UpdateVaultFunction = (input: UpdateVaultInput) => Promise<Vault>
+
+export type GetVaultsFunction = () => Promise<Vault[]>
+
+export type DeleteVaultFunction = (vaultId: string) => Promise<void>
+
+export type VaultsStorage = {
+  getVaults: GetVaultsFunction
+  createVault: CreateVaultFunction
+  updateVault: UpdateVaultFunction
+  deleteVault: DeleteVaultFunction
+}
 
 type MergeVaultsWithCoinsInput = {
   vaults: Vault[]
